DBZ-673 xid must be long in wal2jon decoder

+++ debezium-connector-postgres/src/main/java/io/debezium/connector/postgresql/connection/ReplicationMessage.java
@@ -63,7 +63,7 @@ public interface ReplicationMessage {
     /**
      * @return An id of transaction to which this change belongs
      */
-    public int getTransactionId();
+    public long getTransactionId();
 
     /**
      * @return Table changed
